Address PR review feedback
Consolidate bank-trustees Azure Function into banks
function file, matching the pattern used by
bankruptcy-software. Fix mock totalPages to match
server contract (1 not 0 for empty results). Remove
unnecessary retrievePage pass-through wrapper.

Jira ticket: CAMS-752
