--- backend/function-apps/api/bank-trustees/bank-trustees.function.ts
@@ -1,35 +1,0 @@
-import { app, HttpRequest, HttpResponseInit, InvocationContext } from '@azure/functions';
-import ContextCreator from '../../azure/application-context-creator';
-import { toAzureError, toAzureSuccess } from '../../azure/functions';
-import { BankTrusteesController } from '../../../lib/controllers/bank-trustees/bank-trustees.controller';
-
-const MODULE_NAME = 'BANK-TRUSTEES-FUNCTION';
-
-export default async function handler(
-  request: HttpRequest,
-  invocationContext: InvocationContext,
-): Promise<HttpResponseInit> {
-  const logger = ContextCreator.getLogger(invocationContext);
-
-  try {
-    const context = await ContextCreator.applicationContextCreator({
-      invocationContext,
-      logger,
-      request,
-    });
-
-    context.session = await ContextCreator.getApplicationContextSession(context);
-    const controller = new BankTrusteesController(context);
-    const response = await controller.handleRequest(context);
-    return toAzureSuccess(response);
-  } catch (error) {
-    return toAzureError(logger, MODULE_NAME, error);
-  }
-}
-
-app.http('bank-trustees', {
-  methods: ['GET'],
-  authLevel: 'anonymous',
-  handler,
-  route: 'banks/{bankId}/trustees',
-});
